requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.25
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.4
django-allauth==0.20.0
django-celery==3.2.1
django-encrypted-fields==1.1.2
django-mail-templated==2.6.5
dnspython==1.15.0
email-validator==1.0.2
factory-boy==2.9.2
Faker==0.8.1
funcsigs==1.0.2
idna==2.6
ipaddress==1.0.18
kombu==3.0.37
mock==2.0.0
oauthlib==2.0.2
pbr==3.1.1
psycopg2==2.7.3.1
pyasn1==0.3.3
pycrypto==2.6.1
python-dateutil==2.6.1
python-keyczar==0.716
python-openid==2.2.5
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
six==1.10.0
UkPostcodeParser==1.1.1
urllib3==1.22
